Add EOY balance validation for validation of user input
Added support for user-provided EOY balance (cash) validation to the tax
reporting endpoint. A new optional object is now passed to the tax
reporting endpoint which will handler cash balances for the last two
years. If there is a mismatch of less than 10 USD, we will to a cash
adjustment and log a warning that adjustment has been performed. If the
difference in expected cash vs. calculated cash is greater than 10 USD,
we will not perform a cash adjustment and an error is logged.